+++ ds2t/mateus_alves_da_silva/contatos.js
@@ -1,0 +1,186 @@
+export const contatos = [
+  {
+    "name": "Ana Maria",
+    "description": "Frontend Developer",
+    "image": "img/contact1.png",
+    "messages": [
+      {
+        "sender": "me",
+        "content": "Hello Leonid, how are you doing today?",
+        "time": "14:20"
+      },
+      {
+        "sender": "Ana Maria",
+        "content": "Hi, I'm doing fine. Thanks for asking.",
+        "time": "14:25"
+      },
+      {
+        "sender": "me",
+        "content": "Great to hear that. Do you have any plans for the weekend?",
+        "time": "14:30"
+      },
+      {
+        "sender": "Ana Maria",
+        "content": "Not yet. I'm thinking about going to the beach. What about you?",
+        "time": "14:35"
+      }
+    ]
+  },
+  {
+    "name": "Jane Smith",
+    "description": "UI Designer",
+    "image": "img/contact2.png",
+    "messages": [
+      {
+        "sender": "me",
+        "content": "Hi Jane, how's it going?",
+        "time": "09:00"
+      },
+      {
+        "sender": "Jane Smith",
+        "content": "I'm good, thanks. How about you?",
+        "time": "09:05"
+      },
+      {
+        "sender": "me",
+        "content": "I'm doing well. By the way, have you seen the latest designs for the new project?",
+        "time": "09:10"
+      },
+      {
+        "sender": "Jane Smith",
+        "content": "Yes, I have. They look great!",
+        "time": "09:15"
+      }
+    ]
+  },
+  {
+    "name": "Mark Johnson",
+    "description": "Backend Developer",
+    "image": "img/contact3.png",
+    "messages": [
+      {
+        "sender": "me",
+        "content": "Hi Mark, how's everything going?",
+        "time": "13:00"
+      },
+      {
+        "sender": "Mark Johnson",
+        "content": "Everything's going well. How about you?",
+        "time": "13:05"
+      },
+      {
+        "sender": "me",
+        "content": "I'm doing fine. By the way, do you have any updates on the API development?",
+        "time": "13:10"
+      },
+      {
+        "sender": "Mark Johnson",
+        "content": "Yes, we're making good progress. We should be able to deliver it on schedule.",
+        "time": "13:15"
+      }
+    ]
+  },
+  {
+    "name": "Sarah Lee",
+    "description": "Software Engineer",
+    "image": "img/contact4.png",
+    "messages": [
+      {
+        "sender": "me",
+        "content": "Hey Sarah, como estão as coisas?",
+        "time": "10:00"
+      },
+      {
+        "sender": "Sarah Lee",
+        "content": "Ei! Tudo bem por aqui e contigo?",
+        "time": "10:05"
+      },
+      {
+        "sender": "me",
+        "content": "Estou bem também! O que tem feito ultimamente?",
+        "time": "10:10"
+      },
+      {
+        "sender": "Sarah Lee",
+        "content": "Tenho trabalhado bastante, mas está tudo bem. E você?",
+        "time": "10:15"
+      }
+    ]
+  },
+  {
+    "name": "John Doe",
+    "description": "Designer",
+    "image": "img/contact5.png",
+    "messages": [
+      {
+        "sender": "John Doe",
+        "content": "Oi, tudo bem?",
+        "time": "11:00"
+      },
+      {
+        "sender": "me",
+        "content": "Tudo ótimo! E contigo?",
+        "time": "11:05"
+      },
+      {
+        "sender": "John Doe",
+        "content": "Estou bem também! Viu aquela nova ferramenta de design?",
+        "time": "11:10"
+      },
+      {
+        "sender": "John Doe",
+        "content": "Acho que você vai gostar!",
+        "time": "11:10"
+      }
+    ]
+  },
+  {
+    "name": "Julia Smith",
+    "description": "Product Manager",
+    "image": "img/contact6.png",
+    "messages": [
+      {
+        "sender": "Julia Smith",
+        "content": "Oi, como está indo o projeto?",
+        "time": "12:00"
+      },
+      {
+        "sender": "me",
+        "content": "Está indo bem, estamos quase terminando a fase de testes!",
+        "time": "12:05"
+      },
+      {
+        "sender": "Julia Smith",
+        "content": "Ótimo! E como está o time?",
+        "time": "12:10"
+      },
+      {
+        "sender": "me",
+        "content": "Todos estão trabalhando duro e focados no objetivo!",
+        "time": "12:15"
+      }
+    ]
+  },
+  {
+    "name": "Max Kellerman",
+    "description": "Marketing Manager",
+    "image": "img/contact7.png",
+    "messages": [
+      {
+        "sender": "Max Kellerman",
+        "content": "Ei! Vi que vocês estão com uma nova campanha no ar!",
+        "time": "13:00"
+      },
+      {
+        "sender": "me",
+        "content": "Sim, estamos testando algumas coisas novas!",
+        "time": "13:05"
+      },
+      {
+        "sender": "Max Kellerman",
+        "content": "Bom trabalho, fico feliz em ver vocês tentando coisas novas!",
+        "time": "13:10"
+      }
+    ]
+  }
+]
