+++ ds2t/mateus_alves_da_silva/css/reset.css
@@ -1,0 +1,13 @@
+* {
+    padding: 0;
+    margin: 0;
+    box-sizing: border-box;
+}
+
+ul, ol, li {
+    list-style: none;
+}
+a{
+    text-decoration: none;
+    color: inherit;
+}
