+++ ds2t/mateus_alves_da_silva/index.html
@@ -3,9 +3,69 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Aplicativo</title>
+    <script src="https://kit.fontawesome.com/4f821b9693.js" crossorigin="anonymous"></script>
+    <link rel="stylesheet" href="./css/reset.css">
+    <link rel="stylesheet" href="./css/style.css">
+    <script src="index.js" defer type="module"></script>
+    <title>WhatsApp</title>
 </head>
 <body>
-    
+    <div class="container">
+        <div class="container_contatos">
+            <div class="header_contatos">
+                <div class="imagem_perfil">
+                    <img src="./img/fotoPerfil.png" alt="foto de perfil">
+                </div>
+                <ul class="navegacao_perfil">
+                    <li>
+                        <i class="fa-solid fa-users"></i>
+                    </li>
+                    <li>
+                        <i class="fas fa-circle-notch"></i>
+                    </li>
+                    <li>
+                        <i class="fa-solid fa-message"></i>
+                    </li>
+                    <li>
+                        <i class="fa-solid fa-ellipsis-vertical"></i>
+                    </li>
+                </ul>
+            </div>
+            <div class="barra_pesquisa">
+                <div class="pesquisa">
+                    <i class="fa-solid fa-magnifying-glass"></i>
+                    <input type="text" placeholder="Pesquisar ou começar uma nova conversa" class="caixa_pesquisa">
+                </div>
+                <div class="filtro">
+                    <i class="fa-solid fa-filter"></i>
+                </div>
+            </div>
+            
+            <ul class="chat_list" id="contatoCard" ></ul>
+        </div>
+
+        <div class="container_mensagens">
+            <div class="header_mensagens">
+                <div class="container_header">
+                    <div class="imagem_perfil">
+                        <img src="./img/fotoPerfil.png" alt="">
+                    </div>
+                    <div class="container_perfil">
+                        <span class="info_nome">Mateus Alves</span>
+                        <span class="info_conversa">online</span>
+                    </div>
+                </div>
+                <ul class="nav_perfil_direito">
+                    <li>
+                        <i class="fa-solid fa-magnifying-glass"></i>
+                    </li>
+                    <li>
+                        <i class="fa-solid fa-ellipsis-vertical"></i>
+                    </li>
+                </ul>
+            </div>
+        </div>
+
+    </div>
 </body>
 </html>
